Move useWindowDimensions hook to a global folder
